test(frontend): share mock backend test setup
